requirements.txt
pandas
streamlit
biopython==1.81